requirements.txt
streamlit
groq
pdfplumber
python-docx
python-pptx
fpdf
python-pptx
